requirements/base.txt
Django==1.8.4
Pillow==3.0.0
argparse==1.4.0
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-storages-redux==1.3
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
defusedxml==0.4.1
django-allauth==0.23.0
oauthlib==1.0.3
python3-openid==3.0.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
static3==0.6.1
wheel==0.24.0
